Fix issue #10: implement Lazy() decorator. This allows deferred evaluation of Joi schemas, which allows you to validate recursive data structures.
The current syntax isn't ideal. The callback must return a Joi schema, which means you need to call `getJoiSchema()` on your validation class. If the property is an array, you also need to wrap it in something like `Joi.array().items()`. This ends up quite verbose, so it'd be nice to implement a more concise API at some point.