Fix allowing type unions, as long as the required type is at least one member of the union. This is useful for things like `myProperty: string | null;` or `myProperty?: number;`.
I worked out the type through sheer force of willpower, hopefully it's right. The intention is to create a type, that has the given property (e.g. `myProperty`), with either the property's original type (`string | null`) if it is assignable to a given desired type (`string`), or if it's the wrong type, the desired type (`number`). This allows union types and gives meaningful compiler error messages.